+++ src/interfaces/plugins/noDataRenderer/noDataRenderer.interface.ts
@@ -1,4 +1,5 @@
 import {TemplateRef} from '@angular/core';
+import {NgIfContext} from '@angular/common';
 
 import {VisualPluginOptions} from '../../visualPluginOptions/visualPluginOptions.interface';
 import {GridPlugin} from '../../gridPlugin/gridPlugin.interface';
@@ -37,7 +38,7 @@ export interface NoDataRendererOptions<TCssClasses = unknown> extends VisualPlug
     /**
      * Custom template for no data
      */
-    template: TemplateRef<unknown>;
+    template: TemplateRef<NgIfContext<boolean>>|null;
 }
 
 /**