+++ src/plugins/contentRenderer/contentRenderer.interface.ts
@@ -108,18 +108,3 @@ export interface BasicOrderableColumn<TData = any> extends BasicGridColumn<TData
     orderingClass: string;
 }
 
-/**
- * Definition of simple ordering using one column
- */
-export interface SimpleOrdering
-{
-    /**
-     * Order by column name
-     */
-    orderBy?: string;
-
-    /**
-     * Order by column direction
-     */
-    orderByDirection: OrderByDirection;
-}