+++ src/plugins/noDataRenderer/index.ts
@@ -1,0 +1,2 @@
+export * from './simple/simpleNoDataRenderer.component';
+export * from './simple/simpleNoDataRenderer.interface';