+++ src/plugins/rowSelector/components.ts
@@ -1,2 +1,0 @@
-export * from './basic/basicRowSelector.component';
-export * from './limited/limitedRowSelector.component';