+++ src/plugins/rowSelector/index.ts
@@ -1,3 +1,0 @@
-export * from './rowSelector.interface';
-export * from './basic/basicRowSelector.interface';
-export * from './limited/limitedRowSelector.interface';